# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Topping CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List all meetings
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/check-conflict:
    post:
      operationId: checkConflict
      tags: [meetings]
      summary: Check for scheduling conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConflictCheckInput"
      responses:
        "200":
          description: Conflict check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictCheckResult"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    patch:
      operationId: updateMeeting
      tags: [meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /meetings/team-calendar:
    get:
      operationId: getTeamCalendar
      tags: [meetings]
      summary: Get team calendar view
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: string
        - name: end
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Team calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  # ─── TASKS ──────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── KPI ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Meeting:
      type: object
      required: [id, title, startTime, endTime, status]
      properties:
        id:
          type: integer
        title:
          type: string
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        onlineLink:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show, follow_up]
        outcome:
          type: ["string", "null"]
          enum: [won, lost, proposal_sent, follow_up_required, null]
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/MeetingAttendee"
        dealId:
          type: ["integer", "null"]
        leadId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MeetingAttendee:
      type: object
      required: [userId, name]
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
    MeetingInput:
      type: object
      required: [title, startTime, endTime]
      properties:
        title:
          type: string
        clientName:
          type: string
        companyName:
          type: string
        location:
          type: string
        onlineLink:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        attendeeIds:
          type: array
          items:
            type: integer
        dealId:
          type: integer
        leadId:
          type: integer
        clientId:
          type: integer
        ownerId:
          type: integer
        notes:
          type: string
    MeetingUpdate:
      type: object
      properties:
        title:
          type: string
        clientName:
          type: string
        companyName:
          type: string
        location:
          type: string
        onlineLink:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
        outcome:
          type: string
        attendeeIds:
          type: array
          items:
            type: integer
        dealId:
          type: integer
        leadId:
          type: integer
        clientId:
          type: integer
        notes:
          type: string
    ConflictCheckInput:
      type: object
      required: [startTime, endTime, attendeeIds]
      properties:
        startTime:
          type: string
        endTime:
          type: string
        attendeeIds:
          type: array
          items:
            type: integer
        excludeMeetingId:
          type: integer
    ConflictCheckResult:
      type: object
      required: [hasConflict]
      properties:
        hasConflict:
          type: boolean
        conflicts:
          type: array
          items:
            type: object
            properties:
              userId:
                type: integer
              userName:
                type: string
              conflictingMeetingId:
                type: integer
              conflictingMeetingTitle:
                type: string
              start:
                type: string
              end:
                type: string

    # ─── TASK ─────────────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, title, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        dealId:
          type: ["integer", "null"]
        meetingId:
          type: ["integer", "null"]
        responseNote:
          type: ["string", "null"]
          description: "Assignee's reply: progress update or reason the task isn't done"
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        dueDate:
          type: string
        assigneeId:
          type: integer
        dealId:
          type: integer
        meetingId:
          type: integer
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
        assigneeId:
          type: integer
        responseNote:
          type: ["string", "null"]

    # ─── KPI ──────────────────────────────────────────────────────────────────
...